move storage under storages, create a new example for cache usage
