+++ examples/pytorch/graphsage/advanced/train_sampling_unsupervised_cache.py
@@ -73,7 +73,10 @@ def run(proc_id, n_gpus, args, devices, data):
     if args.data_device == 'gpu':
         nfeat = nfeat.to(device)
     elif args.data_device == 'uva':
-        nfeat = dgl.contrib.UnifiedTensor(nfeat, device=device)
+        if args.cache_size <= 0:
+            nfeat = dgl.contrib.UnifiedTensor(nfeat, device=device)
+        else:
+            nfeat = dgl.contrib.GPUCachedTensor(nfeat, args.cache_size)
     in_feats = nfeat.shape[1]
 
     # Create PyTorch DataLoader for constructing blocks
@@ -135,6 +138,7 @@ def run(proc_id, n_gpus, args, devices, data):
         for step, (input_nodes, pos_graph, neg_graph, blocks) in enumerate(dataloader):
             input_nodes = input_nodes.to(nfeat.device)
             batch_inputs = nfeat[input_nodes].to(device)
+            cache_hit_rate = nfeat.hit_rate if hasattr(nfeat, 'hit_rate') else 0
             blocks = [block.int() for block in blocks]
             d_step = time.time()
 
@@ -154,8 +158,8 @@ def run(proc_id, n_gpus, args, devices, data):
             iter_t.append(t - d_step)
             if step % args.log_every == 0 and proc_id == 0:
                 gpu_mem_alloc = th.cuda.max_memory_allocated() / 1000000 if th.cuda.is_available() else 0
-                print('[{}]Epoch {:05d} | Step {:05d} | Loss {:.4f} | Speed (samples/sec) {:.4f}|{:.4f} | Load {:.4f}| train {:.4f} | GPU {:.1f} MB'.format(
-                    proc_id, epoch, step, loss.item(), np.mean(iter_pos[3:]), np.mean(iter_neg[3:]), np.mean(iter_d[3:]), np.mean(iter_t[3:]), gpu_mem_alloc))
+                print('[{}]Epoch {:05d} | Step {:05d} | Loss {:.4f} | Speed (samples/sec) {:.4f}|{:.4f} | Load {:.4f}| train {:.4f} | GPU {:.1f} MB | Hit Rate {:.4f}'.format(
+                    proc_id, epoch, step, loss.item(), np.mean(iter_pos[3:]), np.mean(iter_neg[3:]), np.mean(iter_d[3:]), np.mean(iter_t[3:]), gpu_mem_alloc, cache_hit_rate))
             tic_step = time.time()
 
         toc = time.time()
@@ -250,6 +254,7 @@ if __name__ == '__main__':
                                 "Use 'cpu' to keep the features on host memory and "
                                 "'uva' to enable UnifiedTensor (GPU zero-copy access on "
                                 "pinned host memory).")
+    argparser.add_argument('--cache-size', type=int, default=0)
     args = argparser.parse_args()
 
     main(args)
