+++ python/dgl/contrib/__init__.py
@@ -4,4 +4,4 @@ from . import dist_sampling
 from .dis_kvstore import KVClient, KVServer
 from .dis_kvstore import read_ip_config
 from .unified_tensor import UnifiedTensor
-from .gpu_cache import GPUCache, GPUCachedTensorStorage, GPUCachedTensor
+from .gpu_cached_tensor import GPUCachedTensor
